added /fight
