+++ src/slash.py
@@ -1,3 +1,4 @@
+#!/bin/python3
 # Links
 # https://docs.pycord.dev/en/master/api.html
 # https://guide.pycord.dev/interactions/application-commands/slash-commands/
@@ -6,11 +7,12 @@
 import os
 import discord
 from dotenv import load_dotenv
-import re
 
 # Local Imports
 from modules.dice import dice
 from modules.dice import ob_dice
+from modules.functions import prettifyDice
+from modules.functions import splitRollString
 
 # Variables from .env
 load_dotenv()
@@ -35,65 +37,6 @@ elif DEBUG_GUILDS is not None:
     bot = discord.Bot(debug_guilds=[DEBUG_GUILDS])
 
 
-# Functions
-def prettifyDice(rolls):
-    out = []
-    for i in rolls:
-        if i == "1":
-            emoji = ":one:"
-        elif i == "2":
-            emoji = ":two:"
-        elif i == "3":
-            emoji = ":three:"
-        elif i == "4":
-            emoji = ":four:"
-        elif i == "5":
-            emoji = ":five:"
-        elif i == "6":
-            emoji = ":six:"
-        else:
-            emoji = i
-        out.append(str(emoji))
-    return ', '.join(out)
-
-
-def splitRollString(roll: str, rollType: str, DEBUG):
-    # Validate pattern
-    if DEBUG == "True":
-        print("\nRegex Debug")
-    regex = r'[0-9]+(T|t|D|d)[0-9]+($|\+[0-9]+$)'
-    if re.match(regex, roll):
-        if DEBUG == "True":
-            print("Regex matched")
-    else:
-        if DEBUG == "True":
-            print("Regex NOT matched")
-        return
-    # Pattern to split: "[(T|t|D|d)+$]\s*"
-    RollSplit = re.split(r"[(T|t|D|d)+$]\s*", roll)
-    number_of_rolls = RollSplit[0]
-    if rollType == "roll":
-        sides_to_die = RollSplit[1]
-    elif rollType == "ob":
-        sides_to_die = 6
-    if len(RollSplit) == 2:
-        bonus = 0
-    elif len(RollSplit) == 3:
-        bonus = RollSplit[2]
-    else:
-        if DEBUG == "True":
-            print("\nlen RollSplit Debug")
-            print("len of RollSplit should only be 2 or 3")
-            print("Len: ", len(RollSplit))
-    # Validate content
-    if DEBUG == "True":
-        print("\nRollSplit Debug")
-        print("Rolls : ", number_of_rolls)
-        print("Sides : ", sides_to_die)
-        print("Bonus : ", bonus)
-    return number_of_rolls, sides_to_die, bonus
-
-
 # Commands
 @bot.command(description="Sends the bot's latency.")
 async def ping(ctx):
@@ -101,7 +44,11 @@ async def ping(ctx):
 
 
 @bot.command(
-    description="Scalable dice, Rolls a die in NdN+bonus or NtN format.")
+    description="""
+    Scalable dice, Rolls a die in NdN+bonus or NtN format.
+    Example: /roll 1d100
+    """
+    )
 async def roll(ctx, roll: discord.Option(str)):
     rollType = "roll"
     try:
@@ -143,8 +90,8 @@ async def roll(ctx, roll: discord.Option(str)):
 
 @bot.command(
     description="""
-    ob dice, Rolls two additional dice for each rolled six.
-    Hardcoded for six sided dice.
+    ob dice, replace sixes with two additional dice.
+    D6 Hardcoded. Example: /ob 3t6+3
     """)
 async def ob(ctx, roll: discord.Option(str)):
     rollType = "ob"
@@ -195,6 +142,7 @@ async def ob(ctx, roll: discord.Option(str)):
         print(sum_rolls)
         print(int(number_of_rolls))
         print(int(number_of_rolls) * 2)
+
     # Determine if its a bad roll.
     # if sum_rolls < (int(number_of_rolls) * 2):
     #     # React
@@ -203,7 +151,63 @@ async def ob(ctx, roll: discord.Option(str)):
     #     # React
     #     await ctx.message.add_reaction(emoji_die)
     #     await ctx.message.add_reaction(emoji_roll)
+
     # Send results
+    await ctx.respond(results)
+
+
+@bot.command(
+    description="""
+    Roll for Fight, rolls ob + t100 for hit.
+    Example: /fight 2t6+2
+    """
+    )
+async def fight(ctx, roll: discord.Option(str)):
+    rollType = "ob"
+    try:
+        # Split roll to vars
+        number_of_rolls, sides_to_die, bonus = splitRollString(
+            roll,
+            rollType,
+            DEBUG
+            )
+
+        # Roll the ob/infinite dice
+        ob_sum_rolls, ob_rolls, ob_raw_rolls, ob_sixes, ob_total = ob_dice(
+            int(number_of_rolls),
+            int(bonus)
+            )
+
+        # Roll the t100 die
+
+        d100_sum_rolls, d100_raw_rolls, d100_total = dice(
+            int(1),
+            int(0),
+            int(100)
+            )
+
+        # Make Pretty
+        pretty_rolls = prettifyDice(ob_rolls)
+        # semiPrettyRolls = ', '.join(d100_raw_rolls)
+
+        # Build result string
+        # results = string_rolled + " {ROLL}\n".format(
+        #    ROLL=roll.upper()
+        #    )
+        results = "OB Rolls: {ROLLS}".format(
+            ROLLS=pretty_rolls
+            )
+        if int(bonus) != 0:
+            results = results + "+ {BONUS}\n".format(BONUS=bonus)
+        else:
+            results = results + "\n"
+        results = results + "OB Total: {TOTAL}\n".format(TOTAL=ob_total)
+        results = results + "d100 Total: {TOTAL}\n".format(TOTAL=d100_total)
+    except ValueError:
+        await ctx.respond("Format has to be in NtN+N, NtN, NdN+N or NdN")
+        return
+
+    # send results
     await ctx.respond(results)
